backend/requirements.txt
gitpython>=3.1.40
pylint>=3.0.2
pytest>=7.4.3
PyGithub>=2.1.1
openai>=1.12.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0 